requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
pydantic[email]==2.5.0
openai==1.3.7
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.0
ollama
rapidocr-onnxruntime
numpy
tensorflow
pymupdf
rapidocr-onnxruntime
python-telegram-bot
vobject
google-generativeai
requests
google-genai